(E008) Consumer diagnostics.

+++ src/NKafka/Client/Consumer/KafkaMessagePackage.cs
@@ -6,12 +6,12 @@ namespace NKafka.Client.Consumer
     [PublicAPI]
     public sealed class KafkaMessagePackage
     {
-        public readonly int PackageNumber;
+        public readonly long PackageNumber;
 
         [NotNull, ItemNotNull]
         public readonly IReadOnlyList<KafkaMessage> Messages;
 
-        public KafkaMessagePackage(int packageNumber, [NotNull, ItemNotNull] IReadOnlyList<KafkaMessage> messages)
+        public KafkaMessagePackage(long packageNumber, [NotNull, ItemNotNull] IReadOnlyList<KafkaMessage> messages)
         {
             Messages = messages;
             PackageNumber = packageNumber;
@@ -21,12 +21,12 @@ namespace NKafka.Client.Consumer
     [PublicAPI]
     public sealed class KafkaMessagePackage<TKey, TData>
     {
-        public readonly int PackageNumber;
+        public readonly long PackageNumber;
 
         [NotNull, ItemNotNull]
         public readonly IReadOnlyList<KafkaMessage<TKey,TData>> Messages;
 
-        public KafkaMessagePackage(int packageNumber, [NotNull, ItemNotNull] IReadOnlyList<KafkaMessage<TKey, TData>> messages)
+        public KafkaMessagePackage(long packageNumber, [NotNull, ItemNotNull] IReadOnlyList<KafkaMessage<TKey, TData>> messages)
         {
             Messages = messages;
             PackageNumber = packageNumber;
